Move LocaleMapping
